% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFibre.r
\name{FindFibre}
\alias{FindFibre}
\title{Find fibre}
\usage{
FindFibre(A, y)
}
\arguments{
\item{A}{A configuration matrix}

\item{y}{Vector of observed counts}
}
\value{
Matrix with columns giving fibre elements
}
\description{
Computes solution set (fibre) for y=Ax for non-negative integers
}
