% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeBasis.r
\name{LatticeBasis}
\alias{LatticeBasis}
\title{Computes partition lattice basis}
\usage{
LatticeBasis(A, mu = NULL, reorder = T)
}
\arguments{
\item{A}{A configuration matrix}

\item{mu}{Optional specification of E\link{x} promote favourable ordering of colums of A (i.e. entries of x)}

\item{reorder}{All columns of A to be reordered to compute basis? Output still matches original column ordering}
}
\value{
List containing partition lattice basis U, determinant of A1 in A = \link{A1, A2}, column ordering
}
\description{
Computes solution set (fibre) for y=Ax for non-negative integers
}
