% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEdgeLoading.r
\name{MaxEdgeLoading}
\alias{MaxEdgeLoading}
\title{Maximum edge loading}
\usage{
MaxEdgeLoading(
  A,
  U,
  y,
  U.reorder = 1:nrow(U),
  lambda = NULL,
  Fibre = NULL,
  UNIFORM = TRUE
)
}
\arguments{
\item{A}{A configuration matrix}

\item{U}{A paritition lattice basis (given by the columns of the matrix)}

\item{y}{Vector of observed counts}

\item{U.reorder}{Recordering of columns of A/rows of U}

\item{lambda}{Rate parameters, need specifying if using Poisson model}

\item{Fibre}{Fibre (solution set) for y=Ax}

\item{UNIFORM}{Indicator for uniform model (otherwise Poisson)}
}
\value{
List containing loading and identifier for most loaded edge
}
\description{
Computes maximum edge loading for transition graph.
}
