% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLEM.r
\name{SLEM}
\alias{SLEM}
\title{SLEM}
\usage{
SLEM(A, U, y, lambda = NULL, Fibre = NULL, UNIFORM = TRUE)
}
\arguments{
\item{A}{A configuration matrix}

\item{U}{A paritition lattice basis (given by the columns of the matrix)}

\item{y}{Vector of observed counts}

\item{lambda}{Rate parameters, need specifying if using Poisson model}

\item{Fibre}{Fibre (solution set) for y=Ax}

\item{UNIFORM}{Indicator for uniform model (otherwise Poisson)}
}
\value{
List containing SLEM, transition matrix and fibre
}
\description{
Computes second largest eigvenvalue modulus for transition matrix
}
