% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEulerian.r
\name{testEulerian}
\alias{testEulerian}
\title{Tests whether matrix is Eulerian}
\usage{
testEulerian(U, c0 = FALSE, r0 = FALSE, SIGN = T, verbose = FALSE)
}
\arguments{
\item{U}{A lattice basis or matrix of signs for lattice basis}

\item{c0}{Logical - test for c0-Eulerian (Eulerian with zero column sums)}

\item{r0}{Logical - test for r0-Eulerian (Eulerian with zero row sums)}

\item{SIGN}{Logical - is U a matrix of signs? (If false, U is redefined as matrix of signs in function)}

\item{verbose}{Monitor progress? Defaults to false.}
}
\value{
List containing UM (indicator for unimodularity) and d (vector determinants of submatrices checked)
}
\description{
Tests whether matrix is Eulerian, or c0-Eulerian, or cr0-Eulerian
}
