% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUM.r
\name{testUM}
\alias{testUM}
\title{Tests unimodularity for configutation matrix}
\usage{
testUM(A, sim.size = 1e+06, verbose = FALSE)
}
\arguments{
\item{A}{A configuration matrix}

\item{sim.size}{Just check a subset of sim.size submatrices if number of submatrices exceeds that number. Defaults to 1e6.}

\item{verbose}{Monitor progress? Defaults to false.}
}
\value{
List containing UM (indicator for unimodularity) and d (vector determinants of submatrices checked)
}
\description{
Tests unimodularity for configutation matrix. For large matrices, checks just a sample of submatrices.
}
